+++ crates/outlayer/sdk/src/lib.rs
@@ -1,23 +1,4 @@
-#[cfg(target_family = "wasm")]
-mod sys;
+mod host;
 
-use std::sync::LazyLock;
-
-pub use defuse_outlayer_host as host;
-
-#[cfg(target_family = "wasm")]
-type Host = sys::SysHost;
-
-#[cfg(not(target_family = "wasm"))]
-type Host = defuse_outlayer_worker_host::WorkerHost;
-
-pub static HOST: LazyLock<Host> = LazyLock::new(|| {
-    #[cfg(target_family = "wasm")]
-    {
-        sys::SysHost
-    }
-    #[cfg(not(target_family = "wasm"))]
-    {
-        unimplemented!("init_host is not implemented");
-    }
-});
+pub use defuse_outlayer_host::*;
+pub use host::HOST;
